requirements.txt
pynput==1.7.3
sqlalchemy==1.4.30
Pillow==8.3.2
opencv-python==4.5.3.56
